SMD-842 - Moving 'table_configs' under 'tables'
